Fix https://github.com/angelozerr/lsp4xml/issues/386
This PR fixes https://github.com/angelozerr/lsp4xml/issues/386 The
problem comes from of populate of DTD element children when it is
loaded. As elements can be declared in many area, we need to check that
children element name was not added.

This PR is just to see how to fix problem. Please ignore it and creates
a new PR with tests.

Signed-off-by: azerr <azerr@redhat.com>