Fix bug with completion on XML Schema elements
Completion on XML Schema like xs:annotation, inside a xsd file is broken. The getDocument method is called in the services (competion, hover, etc ) wher echeck of NPE should be done but it is called too in XMLCatalogURIResolverExtension, and other extension where getDocument should continue to return PR (and not throw a CancellableException). 

This PR is a simple fix to resolve problem with completion on XML Schema.