Add org.eclipse.xtend.lib as it is no longer a dependency of lsp4j
Previously org.eclipse.lsp4j transitively required org.eclipse.xtend.lib
but this is no longer the case, this adds it back here, but as lsp4j can
work without it lemminx should also consider to drop this requirement.

Also the API for WorkspaceFolder has changed here, so a test needs to be
adjusted here.