Add an uber-bundle artifact
Currently lemminx only provides an uber-jar that has no OSGi meta data
at all, because of this for example wild-webdeveloper needs to embed
this jar instead of use one installed in the OSGi framework.

This now adds a new uber-bundle artifact that is at the moment very
similar to uber-jar but contains some OSGi data useful for writing
extensions. This can then be used as a starting point to incrementally
make it more OSGified, e.g. by export/import its dependencies that are
already available as OSGi bundles (gson, lsp4j, ...)

Signed-off-by: Christoph Läubrich <laeubi@laeubi-soft.de>